requirements.txt
ipykernel
networkx==2.8
gcastle==1.0.3
causal-learn==0.1.3.8
seaborn==0.13.2
torch==2.5.1
jax==0.4.31